backend/requirements.txt
amqp==5.0.7
asgiref==3.4.1
attrs==21.2.0
Automat==20.2.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
bs4==0.0.1
celery==5.2.1
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coloredlogs==15.0.1
constantly==15.1.0
cryptography==36.0.1
cssselect==1.1.0
Django==4.0
django-cors-headers==3.10.1
djangorestframework==3.13.1
h2==3.2.0
hpack==3.0.0
humanfriendly==10.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.4.0
itemloaders==1.0.4
jmespath==0.10.0
kombu==5.2.2
lxml==4.7.1
parsel==1.6.0
priority==1.3.0
prompt-toolkit==3.0.24
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pyOpenSSL==21.0.0
pytz==2021.3
queuelib==1.6.2
requests==2.26.0
Scrapy==2.5.1
scrapy-djangoitem==1.1.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
Twisted==21.7.0
typing_extensions==4.0.1
urllib3==1.26.7
vine==5.0.0
w3lib==1.22.0
wcwidth==0.2.5
zope.interface==5.4.0
